Refactor question rendering into external class
